[Rename] HomeNoneContent 파일명 Rename (#127)
